requirements.txt
pydantic-ai>=0.1.0
openai>=1.0.0
pinecone>=3.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
httpx>=0.25.0
asyncio>=3.4.3
